Test and publish a multi-release jar
The previously published jars only worked on Java 21 because classes that need `--enable-preview` are not compatible between Java releases (even when the API is identical). We now still build the base contents of the jar on Java 21, but we add Java 22 overrides for all classes that have the preview flag set.